+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - **INTERNAL**: Use `str` in enums.
 
+### :bug: Fixed
+
+- Do not mutate `pattern` keywords in response schema validation. [#2749](https://github.com/schemathesis/schemathesis/issues/2749)
+
 ## [4.0.0-alpha.7](https://github.com/schemathesis/schemathesis/compare/v4.0.0-alpha.6...v4.0.0-alpha.7) - 2025-03-21
 
 ### :rocket: Added